+++ changelog.md
@@ -2,10 +2,11 @@
 
 # 3.0.1
 
-- Support for reloading office files after changes.
+- Support for reloading workspace docx after file changes.
 - PDF:
   - Fixed sometimes opening PDF failed.
   - Do not display the sidebar on small screens.
+  - Support export markdown to pdf without outline.
 
 # 3.0.0 2023-3-29